+++ src/UMS.API/Controller/CountryController.cs
@@ -31,7 +31,7 @@ namespace UMS.API.Controller
         }
 
         [HttpGet("get")]
-        public async ValueTask<IActionResult> GetByIdAsync(int id)
+        public async ValueTask<IActionResult> GetByIdAsync(long id)
         {
             var country = await repository.GetByIdAsync(id);
